Moved generation to seperate class
